eval/requirements_prime.txt
absl-py==2.1.0
accelerate==1.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.11.0
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
cmake==3.30.5
compressed-tensors==0.6.0
contourpy==1.3.0
cycler==0.12.1
datasets==3.1.0
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
filelock==3.16.1
fire==0.7.0
fonttools==4.54.1
frozenlist==1.5.0
fschat==0.2.36
fsspec==2024.9.0
func_timeout==4.3.5
gguf==0.10.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
immutabledict==4.2.0
importlib_metadata==8.5.0
interegular==0.3.3
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langdetect==1.0.9
lark==1.2.2
latex2mathml==3.77.0
llvmlite==0.43.0
lm-format-enforcer==0.10.6
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.18
ninja==1.11.1.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
openai==1.57.0
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
outlines==0.0.46
packaging==24.1
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pexpect==4.9.0
pillow==10.4.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.0.0
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
ray==2.38.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.2
svgwrite==1.4.3
sympy==1.12
termcolor==2.5.0
thefuzz==0.22.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vllm-flash-attn==2.6.2
watchfiles==0.24.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==13.1
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.17.1
zipp==3.21.0